requirements.txt
jupyter
openai
SpeechRecognition
azure-cognitiveservices-speech
gtts
TTS
cTTS